refactor dragndrop to Draggable wrapper
