+++ src/main/scala/control/SharedAxisCharts.scala
@@ -40,7 +40,7 @@ class SharedAxisCharts(val dataItems: mutable.HashMap[String, BTreeSeries]) exte
         }
         val chart = BTreeChart.fromTree(this, bTreeData)
         setVgrow(chart, Priority.Always)
-        this.items.add(chart)
+        this.items.add(new DraggableBTreeChartWrapper(chart))
       }
       case _ =>  // shouldn't get here
     }
